Move loadAllResource into care-team package
