+++ packages/fhir-care-team/src/components/CreateEditCareTeam/utils.tsx
@@ -7,6 +7,8 @@ import { FHIR_CARE_TEAM, URL_CARE_TEAM } from '../../constants';
 import { Dictionary } from '@onaio/utils';
 import { IfhirR4 } from '@smile-cdr/fhirts';
 import { Fields, FormFields } from './Form';
+import { IBundle } from '@smile-cdr/fhirts/dist/FHIR-R4/interfaces/IBundle';
+import { URLParams } from 'opensrp-server-service/dist/types';
 
 export const submitForm = async (
   values: FormFields,
@@ -94,3 +96,30 @@ export function getPatientName(patient: Dictionary | undefined) {
     .filter(Boolean)
     .join(' ');
 }
+
+// TODO - duplicate in #896
+/**
+ * fetch all resources for a certain endpoint
+ *
+ * @param baseUrl - the fhir server url
+ * @param resourceType - the resource type
+ * @param extraFilters - extra filters
+ */
+export const loadAllResources = async (
+  baseUrl: string,
+  resourceType: string,
+  extraFilters: URLParams = {}
+) => {
+  // first get total
+  const summaryFilters = {
+    _summary: 'count',
+    ...extraFilters,
+  };
+  const summary = await new FHIRServiceClass<IBundle>(baseUrl, resourceType).list(summaryFilters);
+  const { total } = summary;
+  const fetchAllFilter = {
+    _count: total,
+    ...extraFilters,
+  };
+  return new FHIRServiceClass<IBundle>(baseUrl, resourceType).list(fetchAllFilter);
+};
